TOTS-80 Added routing to TOTS.
