+++ app/client/package.json
@@ -42,6 +42,7 @@
     "react-color": "2.19.3",
     "react-dom": "18.2.0",
     "react-dropzone": "14.2.3",
+    "react-router-dom": "6.25.1",
     "react-scripts": "5.0.1",
     "react-select": "5.8.0",
     "react-switch": "7.0.0",
