+++ app/client/src/routes/404.ts
@@ -1,0 +1,23 @@
+import { useEffect } from 'react';
+
+// --- components ---
+function PageNotFound() {
+  // redirect to the server side 400.html page
+  useEffect(() => {
+    const location = window.location;
+
+    let url = `${location.origin}/404.html`;
+    if (location.hostname === 'localhost') {
+      url = `${location.protocol}//${location.hostname}:9090/404.html`;
+    }
+
+    // append the original url for tracking purposes
+    url += `?src=${location.href}`;
+
+    window.location.assign(url);
+  }, []);
+
+  return null;
+}
+
+export default PageNotFound;
