+++ app/client/src/routes/Decon.tsx
@@ -1,0 +1,3 @@
+export default function Tods() {
+  return <p>Placeholder...</p>;
+}
