+++ app/client/src/routes/Sampling.tsx
@@ -1,0 +1,5 @@
+import App from 'components/App';
+
+export default function Home() {
+  return <App />;
+}
